requirements.txt
Django>=5.0
gunicorn
whitenoise
python-dotenv
dj-database-url
